Added safe guards, ask before delete

+++ src/github.com/mickep76/etcd-delete/etcd-delete.go
@@ -18,6 +18,7 @@ func main() {
 
 	// Options.
 	version := flag.Bool("version", false, "Version")
+	force := flag.Bool("force", false, "Force delete without asking")
 	node := flag.String("node", "", "Etcd node")
 	port := flag.String("port", "2379", "Etcd port")
 	dir := flag.String("dir", "/", "Etcd directory")
@@ -40,7 +41,17 @@ func main() {
 	}
 	client := etcd.NewClient(conn)
 
+	if !*force {
+		fmt.Printf("Remove path: %s? [yes|no]", strings.TrimRight(*dir, "/"))
+		var query string
+		fmt.Scanln(&query)
+		if query != "yes" {
+			os.Exit(0)
+		}
+	}
+
 	if _, err := client.Delete(strings.TrimRight(*dir, "/"), true); err != nil {
 		log.Fatalf(err.Error())
 	}
+	log.Printf("Removed path: %s", strings.TrimRight(*dir, "/"))
 }
