+++ src/github.com/mickep76/etcd-import/etcd-import.go
@@ -25,6 +25,7 @@ func main() {
 
 	// Options.
 	version := flag.Bool("version", false, "Version")
+	force := flag.Bool("force", false, "Force delete without asking")
 	delete := flag.Bool("delete", false, "Delete entry before import")
 	node := flag.String("node", "", "Etcd node")
 	port := flag.String("port", "2379", "Etcd port")
@@ -76,9 +77,19 @@ func main() {
 	}
 
 	if *delete {
+		if !*force {
+			fmt.Printf("Remove path: %s? [yes|no]", strings.TrimRight(*dir, "/"))
+			var query string
+			fmt.Scanln(&query)
+			if query != "yes" {
+				os.Exit(0)
+			}
+		}
+
 		if _, err := client.Delete(strings.TrimRight(*dir, "/"), true); err != nil {
 			log.Fatalf(err.Error())
 		}
+		log.Printf("Removed path: %s", strings.TrimRight(*dir, "/"))
 	}
 
 	if err = etcdmap.Create(client, strings.TrimRight(*dir, "/"), reflect.ValueOf(m)); err != nil {
